backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4.0
channels>=4.0
daphne>=4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
dj-database-url>=2.1.0
